requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
sqlalchemy>=2.0
PyJWT>=2.8
pydantic-settings>=2.0
python-dotenv>=1.0
psycopg2-binary>=2.9
